Propagate types down to discovered.json
